// libs/database.type.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      profiles: {
        Row: {
          avatar_url: string | null
          email: string
          id: string
          introduce: string | null
          name: string | null
        }
        Insert: {
          avatar_url?: string | null
          email: string
          id: string
          introduce?: string | null
          name?: string | null
        }
        Update: {
          avatar_url?: string | null
          email?: string
          id?: string
          introduce?: string | null
          name?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "profiles_id_fkey"
            columns: ["id"]
            isOneToOne: true
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}